serial16
UTFにたいおうできるかな

+++ droid.js
@@ -22,8 +22,8 @@ var message ="";
         device.open({ stopBits: 0, ctsFlowControl: 0 });
         device.set_receive_handler(function(data) {
             console.log(data);
-            console.log( new Uint8Array(data));
-            message= String.fromCharCode.apply(null,new Uint8Array(data));
+            console.log( new Uint32Array(data));
+            message= String.fromCharCode.apply(null,new Uint32Array(data));
         });
       }
